refactor: 어드민 검색 페이지 위치 변경
